substack-archiver/slack-agent/pyproject.toml
[project]
name = "slack-agent"
version = "0.1.0"
description = "ADK-powered Slack bot over Substack archive — local demo"
requires-python = ">=3.13"
dependencies = [
    "google-adk>=1.0.0",
    "slack-bolt>=1.21.0",
    "slack-sdk>=3.33.0",
    "httpx>=0.28.0",
]
